Added empty If, While and DoWhile referees.
